refactor: Command module classification
