+++ starbot/commands/builtin/__init__.py
@@ -1,0 +1,14 @@
+import starbot.commands.builtin.at.dynamic_at_me
+import starbot.commands.builtin.at.dynamic_at_me_cancel
+import starbot.commands.builtin.at.dynamic_at_me_list
+import starbot.commands.builtin.at.live_on_at_me
+import starbot.commands.builtin.at.live_on_at_me_cancel
+import starbot.commands.builtin.at.live_on_at_me_list
+import starbot.commands.builtin.data.bind
+import starbot.commands.builtin.data.room_data
+import starbot.commands.builtin.data.room_data_total
+import starbot.commands.builtin.data.user_data
+import starbot.commands.builtin.data.user_data_total
+import starbot.commands.builtin.disable_command
+import starbot.commands.builtin.enable_command
+import starbot.commands.builtin.help
